wechat

+++ app/Models/ProjectShareLog.php
@@ -12,9 +12,9 @@ class ProjectShareLog extends Model {
      * @return [type] [description]
      */
     public function createShareLog($project_id){
-        $this->uid = session('wechat_user')['id'];
-        $this->project_id = $project_id;
-        $this->wechat_id = Project::where('id', $project_id)->select('wechat_id');
+        $this->uid = (int)session('wechat_user')['id'];
+        $this->project_id = (int)$project_id;
+        $this->wechat_id = (int)Project::where('id', $project_id)->select('wechat_id');
         $this->save();
     }
 }
